+++ examples/skulpin_app.rs
@@ -2,16 +2,14 @@
 // It's not as flexible as working with winit directly, but it's quick and simple
 
 use skulpin::CoordinateSystem;
-use skulpin::CoordinateSystemHelper;
 use skulpin::LogicalSize;
 use skulpin::skia_safe;
 
 use skulpin::app::AppBuilder;
+use skulpin::app::AppUpdateArgs;
+use skulpin::app::AppDrawArgs;
 use skulpin::app::AppError;
 use skulpin::app::AppHandler;
-use skulpin::app::AppControl;
-use skulpin::app::InputState;
-use skulpin::app::TimeState;
 use skulpin::app::VirtualKeyCode;
 use std::ffi::CString;
 
@@ -54,10 +52,11 @@ impl ExampleApp {
 impl AppHandler for ExampleApp {
     fn update(
         &mut self,
-        app_control: &mut AppControl,
-        input_state: &InputState,
-        _time_state: &TimeState,
+        update_args: AppUpdateArgs,
     ) {
+        let input_state = update_args.input_state;
+        let app_control = update_args.app_control;
+
         if input_state.is_key_down(VirtualKeyCode::Escape) {
             app_control.enqueue_terminate_process();
         }
@@ -65,12 +64,11 @@ impl AppHandler for ExampleApp {
 
     fn draw(
         &mut self,
-        _app_control: &AppControl,
-        _input_state: &InputState,
-        time_state: &TimeState,
-        canvas: &mut skia_safe::Canvas,
-        _coordinate_system_helper: &CoordinateSystemHelper,
+        draw_args: AppDrawArgs,
     ) {
+        let time_state = draw_args.time_state;
+        let canvas = draw_args.canvas;
+
         // Generally would want to clear data every time we draw
         canvas.clear(skia_safe::Color::from_argb(0, 0, 0, 255));