+++ skulpin-app-winit/src/lib.rs
@@ -6,6 +6,8 @@ pub use app::App;
 pub use app::AppHandler;
 pub use app::AppBuilder;
 pub use app::AppError;
+pub use app::AppUpdateArgs;
+pub use app::AppDrawArgs;
 
 mod app_control;
 pub use app_control::AppControl;
@@ -17,6 +19,7 @@ pub use input_state::MouseDragState;
 // These are re-exported winit types
 pub use input_state::VirtualKeyCode;
 pub use input_state::MouseButton;
+pub use input_state::MouseScrollDelta;
 pub use input_state::ElementState;
 pub use input_state::LogicalSize;
 pub use input_state::PhysicalSize;