+++ skulpin-renderer/src/swapchain.rs
@@ -12,6 +12,7 @@ use super::Window;
 
 pub const MAX_FRAMES_IN_FLIGHT: usize = 2;
 
+#[derive(Clone)]
 pub struct SwapchainInfo {
     pub surface_format: vk::SurfaceFormatKHR,
     pub present_mode: vk::PresentModeKHR,